tore out IMC2 - expand for details
I decided to remove IMC2 because it was causing a lot
of warnings, even when the proper defs were enabled.
I decided that the benefit of having IMC2 was not worth
the work it would take to sort out all of the warnings,
because I doubt there are many people out there using
it in this Year of our Lord 2022.

If you want IMC2 back, it still exists in the commits prior
to this one.